+++ src/main/java/com/example/demo/dto/response/LoaiGheResponse.java
@@ -1,10 +1,12 @@
 package com.example.demo.dto.response;
 
+import java.math.BigDecimal;
+
 public class LoaiGheResponse {
 
     private String maLoaiGhe;
     private String tenLoaiGhe;
-    private Double giaPhuThu;
+    private BigDecimal giaPhuThu;
 
     public LoaiGheResponse() {
     }
@@ -25,11 +27,11 @@ public class LoaiGheResponse {
         this.tenLoaiGhe = tenLoaiGhe;
     }
 
-    public Double getGiaPhuThu() {
+    public BigDecimal getGiaPhuThu() {
         return giaPhuThu;
     }
 
-    public void setGiaPhuThu(Double giaPhuThu) {
+    public void setGiaPhuThu(BigDecimal giaPhuThu) {
         this.giaPhuThu = giaPhuThu;
     }
 }
